backend/requirements.txt
fastapi
uvicorn
opencv-python-headless
inference-sdk